renaming, moved testing function to utils file
